requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
python-dotenv==1.0.1
PyJWT==2.8.0
pandas==2.2.1
reportlab==4.1.0
qrcode==7.4.2
Werkzeug==3.0.1
email-validator==2.1.0.post1 